tools/requirements.txt
numpy
pyscreenshot
 # also needs sudo apt-get install imagemagick
Pillow
pynput
opencv-python
